LPD-41245 Properly implement MBeans
