+++ modules/apps/portal-cache/portal-cache-ehcache3-impl/src/main/java/com/liferay/portal/cache/ehcache3/internal/management/CacheMBean.java
@@ -5,41 +5,17 @@
 
 package com.liferay.portal.cache.ehcache3.internal.management;
 
-import org.ehcache.Cache;
-import org.ehcache.config.CacheRuntimeConfiguration;
-
 /**
  * @author Dante Wang
  */
-public class CacheMBean {
+public interface CacheMBean {
 
-	public CacheMBean(String cacheName, Cache<?, ?> cache) {
-		_cacheName = cacheName;
-		_cache = cache;
+	public void clear();
 
-		_cacheRuntimeConfiguration = cache.getRuntimeConfiguration();
-	}
+	public String getKeyType();
 
-	public void clear() {
-		_cache.clear();
-	}
+	public String getName();
 
-	public String getKeyType() {
-		return _cacheRuntimeConfiguration.getKeyType(
-		).getName();
-	}
-
-	public String getName() {
-		return _cacheName;
-	}
-
-	public String getValueType() {
-		return _cacheRuntimeConfiguration.getValueType(
-		).getName();
-	}
-
-	private final Cache<?, ?> _cache;
-	private final String _cacheName;
-	private final CacheRuntimeConfiguration<?, ?> _cacheRuntimeConfiguration;
+	public String getValueType();
 
 }
